move messages library
